readded checking if it is a number

+++ commands/mod commands/clear.js
@@ -7,17 +7,21 @@ module.exports = {
     //run command
     
     num = parseInt(text)
-
-    message.delete()
-    if (num > 0 && num < 101) {
-      message.channel.bulkDelete(text).catch((error) => {
-        if (error.code == 10008) {
-          message.reply(
-            "Cannot clear, as there are messages older than 2 weeks"
-          );
-        }
-        message.channel.send(text+" messages successfully deleted")
-      });
+    
+    if (isNaN("|"+num+"|")) {
+      message.delete()
+      if (num > 0 && num < 101) {
+        message.channel.bulkDelete(text).catch((error) => {
+          if (error.code == 10008) {
+            message.reply(
+              "Cannot clear, as there are messages older than 2 weeks"
+            );
+          }
+          message.channel.send(text+" messages successfully deleted")
+        });
+      } else {
+        message.reply("Please specify a valid number of messages");
+      }
     } else {
       message.reply("Please specify a valid number of messages");
     }
